refactor: LottoNumber 관련 상수 분리
