+++ src/main/kotlin/lotto/domain/LottoNumber.kt
@@ -1,7 +1,5 @@
 package lotto.domain
 
-import lotto.domain.LottoOperator.LOTTO_FIRST_NUMBER
-import lotto.domain.LottoOperator.LOTTO_LAST_NUMBER
 import lotto.exception.InvalidLottoNumberException
 
 @JvmInline
@@ -12,5 +10,12 @@ value class LottoNumber(
         if (value !in LOTTO_FIRST_NUMBER..LOTTO_LAST_NUMBER) {
             throw InvalidLottoNumberException()
         }
+    }
+
+    companion object {
+        const val LOTTO_FIRST_NUMBER = 1
+        const val LOTTO_LAST_NUMBER = 45
+
+        const val LOTTO_SIZE = 6
     }
 }
