--- src/main/kotlin/lotto/domain/LottoOperator.kt
@@ -1,11 +1,0 @@
-package lotto.domain
-
-object LottoOperator {
-    const val LOTTO_FIRST_NUMBER = 1
-    const val LOTTO_LAST_NUMBER = 45
-
-    const val LOTTO_FIRST_INDEX = 0
-    const val LOTTO_LAST_INDEX = 6
-
-    const val LOTTO_SIZE = 6
-}
