+++ src/main/kotlin/lotto/domain/generator/LottoGeneratorStrategy.kt
@@ -4,4 +4,9 @@ import lotto.domain.Lotto
 
 interface LottoGeneratorStrategy {
     fun generate(lottoCount: Int): List<Lotto>
+
+    companion object {
+        const val LOTTO_FIRST_INDEX = 0
+        const val LOTTO_LAST_INDEX = 6
+    }
 }
